Add unit tests for helper functions
